add type for ethereum prop on window object + more clean up on wallet integration
