+++ src/components/Dashboard/VaiInfo.tsx
@@ -82,7 +82,6 @@ function VaiInfo({ settings }: VaiInfoProps) {
         <div className="flex align-center">
           <img src={coinImg} alt="coin" />
           <p>{format(userVaiBalance.dp(2, 1).toString(10))} VAI </p>
-          {/*  @ts-expect-error ts-migrate(2339) FIXME: Property 'ethereum' does not exist on type 'Window... Remove this comment to see the full error message */}
           {(window.ethereum || window.BinanceChain) && (
             <Icon
               className="add-vai-token"
