+++ src/hooks/wallet/useInactiveListener.ts
@@ -6,10 +6,9 @@ export default function useInactiveListener(suppress = false) {
   const { active, error, activate } = useWeb3React();
 
   useEffect(() => {
-    // @ts-expect-error ts-migrate(2339) FIXME: Property 'ethereum' does not exist on type 'Window... Remove this comment to see the full error message
     const { ethereum } = window;
 
-    if (ethereum && ethereum.on && !active && !error && !suppress && !process.env.STORYBOOK) {
+    if (ethereum && 'on' in ethereum && !active && !error && !suppress && !process.env.STORYBOOK) {
       const handleChainChanged = () => {
         // eat errors
         activate(injected, undefined, true).catch(err => {
