+++ src/react-app-env.d.ts
@@ -1,1 +1,6 @@
 // / <reference types="react-scripts" />
+
+interface Window {
+  ethereum?: Record<string, any>;
+  BinanceChain?: unknown;
+}
